Reflect lineNum, side and range to attributes
This is a prerequisite for making the thread widget configurable:
When the thread elements will be created outside of gr-diff and passed
in (to allow different callers using different widgets), we need some
way by which gr-diff can tell where to place these widgets. The plan is
to use these attributes to do that, and require all comment thread
elements passed in via slotting to have these set.

Change-Id: I91162ca2237934c53e6eaba37327e60f1142dfbc

Former-commit-id: 1a0068256dbe6482520a60f7fc89ee46aec52937