Add gr-embed-dashboard element
In Polymer 1.x, the Shadow DOM polyfill requires using the Polymer.dom
API to achieve distribution. There is no way to notify a DOM mutation
happened after the fact. So, when embedding Polymer 1.x elements inside
an app that is not itself a Polymer app, placing arbitrary content
inside a <slot> is not possible.

This element is intended to be a workaround for the duration of
PolyGerrit's existence on Polymer 1.x to allow for slotting of the
<gr-create-change-help> element into the "empty-outgoing" slot of the
<gr-change-list>.

Change-Id: I32a5031df6fff10e3c441ecdf87c7b50d2073e34

Former-commit-id: 3a5019e15ec5a92b0bb854e7c940142a0e3793b9