Fix two tests that were not testing anything
The tests accidentally used assert() instead of assert.equal(). This
still passed because assert is also a function that takes two
parameters, an expression and an error message. The test was passing in
a truthy expected value as the expression, and so it never threw an
error.

In the meanwhile, this test was completely broken and referenced several
no longer existing fields. So I ended up rewriting it to actually test
the methods again.

Change-Id: I173722d92ec4d87a000ac90510de373b2f0e62ea

Former-commit-id: a06d2f38af0655962633da9ff230853969c52a5d