Switch to imperative comment thread creation
We eventually need to move the comment thread creation to gr-diff-host,
and there it cannot be done from a template because we need to hold onto
instances of the thread elements as they are slotted into gr-diff.

I realize that for now this does not look like an improvement, but
having the whole refactoring into one change is to risky, so some of
these intermediate steps look a bit odd. See
https://gerrit-review.googlesource.com/c/gerrit/+/197112
for a WIP CL of where this is going.

Change-Id: If7919bc62ce5c5719ae1c0c682c11f9f73a9c5ec

Former-commit-id: de7141146bbbde12c83e630da3907e6c3ca25f8c