Add metric to measure latency for guessing relevant groups
Guessing groups is invoked from some group backends to find the groups
of a user (e.g. from LdapGroupBackend). The groups of the calling user
are needed for the change etag computation. Hence if guessing groups is
slow this has direct impact on the speed of the change etag computation
which makes it interesting to know how much latency we spend on guessing
relevant groups.

Change-Id: Ibae879a702af007f1a326d01d15553487926bc4a
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 8e07cf65d839eddf17b15f1f84f7fcc158bee271