Fix comment sorting
Returning 0 in a compare function means the two elements are equal, and
in stable sorting implementations their current order stays. The
existing sorting was not symmetric: When called with (no draft, draft),
it would return "equal", but when called with (draft, no draft), it
would return "greater". I think that is a recipe for chaos, and probably
not intended.

This led to tests failing locally but succeeding on CI.

Change-Id: Icb5a097ef89488fb788163f9f5f92a757cde273e

Former-commit-id: 37228643804ebbef8d2cda2aaf1539e290c929e2