Always stub getLoggedIn()
This is to avoid 4xx errors in tests. For these tests, log in status
does not matter, except that if the user is logged in, more calls need
to be stubbed out. So we just set it to always logged out in the
top-level suite, and then set it explicitly to logged out again in the
suite that specifically tests behavior that depends on being logged out.

Change-Id: Idbe0d38e32d64462972a9257cd048d9aab0af085

Former-commit-id: 49ba287b469662b337128c1481722789e18baad2