[CVE-2017-12629] Upgrade Lucene to 5.5.5
This upgrade fixes CVE-2017-12629 [1]. Although this issue only affects
Solr, according to the description, it's probably worth upgrading anyway
since there have been several bug fixes in the intermediate versions.

See the release notes for 5.5.2 [2], 5.5.4 [3] and 5.5.5 [4] for details.

Note: there are no bug fixes listed for 5.5.3.

The only reason we didn't upgrade to the latest version before is because
we had a dependency on Elasticsearch which had a tight coupling with a
specific Lucene version.

[1] https://nvd.nist.gov/vuln/detail/CVE-2017-12629
[2] https://lucene.apache.org/core/5_5_2/changes/Changes.html#v5.5.2.bug_fixes
[3] https://lucene.apache.org/core/5_5_4/changes/Changes.html#v5.5.4.bug_fixes
[4] https://lucene.apache.org/core/5_5_5/changes/Changes.html#v5.5.5.bug_fixes

Bug: Issue 9952
Change-Id: I776e2dc10c86dc6761a0a1ce6644ce5ac384509c

Former-commit-id: fb11ccf5072380899591c05416f557a11cabf6da