Fix Elasticsearch dependency on httpcore-nio
The elasticsearch-rest-client library has an explicit dependency on
version 4.4.5 of httpcore-nio [1], but the version provided by Gerrit
is tied to the same version 4.4.1 as all the other httpcomponents.

Since httpcore-nio is only used by elasticsearch-rest-client, we can
safely provide the required version. We can also restrict visibility
to the elasticsearch package to prevent accidental usage elsewhere.

The same is true for the httpasyncclient component, which we already
provide at the correct version as used by elasticsearch-rest-client,
so also restrict its visibility in the same way.

At the same time, move httpcore-nio and httpasyncclient declarations
up the WORKSPACE file adjacent to the other httpcomponents, and add
a comment to clarify that they are set at explicit versions for ES.

[1] https://search.maven.org/artifact/org.elasticsearch.client/elasticsearch-rest-client/6.4.2/jar

Bug: Issue 9969
Change-Id: Id80f92768667541604df0c53235ea81ecdda9c0e

Former-commit-id: cafa53b95e6d92e8d90e5c7b0c1054c5f47d0007