Refactor ProjectConfig constructors to live in a Factory
In the future, we would like to add a few injected dependencies of
ProjectConfig. The fact that we can't use an assisted factory will
hopefully prevent us from adding too many. Regardless, it's reasonable
to group together the few "read" static constructor method into a single
class along with the bare create method.

Change-Id: I391ae8e37086c22c10e7f342d46e08dd2242df71

Former-commit-id: 0ff5bceaa2d5124e60c094ff173c9680b5453b78