Add delete-project as a core plugin
Also add dependency on Mockito, which is now exported in the
acceptance framework. There are several other plugins using
this dependency, so exporting it in the acceptance framework
will have the nice side effect that we no longer need to keep
updating them individually.

Also note that one of the other plugins that uses Mockito is
the webhooks plugin, which is also planned to be added as a
core plugin.

Feature: Issue 8863
Change-Id: I72c0694ec6fcbabc2ef030014268738aa69cb707

Former-commit-id: 1e150738d8bec72cf22d5a1e8d32ae6af34e8ab5