Don't use List in GerritConfigListener API
Currently GerritConfigListener's configUpdated method returns a List of
ConfigUpdatedEvent.Update.

There are only two types of updates (applied and rejected)
and the API returning lists of updates does not reflect this but
gives the impression that there can be an arbitrary amount
of update types.

Instead of returning a List of ConfigUpdatedEvent.Update that each have
a type, return a Multimap<UpdateResult, ConfigUpdateEntry>.

Removed the ConfigUpdatedEvent.Update class.

Change-Id: I81b4327a5878739087cd8a2ea8c0cff035b6ee57

Former-commit-id: 31300d71e31aa28f502fa01a4ddec5ababd1e9e5