Fix handling comment save in Shadow DOM
e.target in native Shadow DOM will be retargetted to be some containing
element, not the gr-diff-comment element as expected and as it would be
in shady DOM.

I originally wanted to fix this by using Polymer.dom(e).rootTarget but
that has two downsides:
1) It's harder to test, because Polymer.dom does a instanceof check, and
creating and actual Event in the test means we cannot easily set target
because that is readonly.
2) It makes this component more dependent on gr-diff-comment having a
public property comment.

Luckily there is no reason to muck with e.target in the first place
because these event actually have the comment set on the detail aswell,
so just reading it from there circumvents the problem and is easy to
test.

Bug: Issue 10016
Change-Id: I08a546da4ba8d054ff4738f08d1f86ca6e510a3c

Former-commit-id: 75d6058f4e6278af609cc89829a834906c80db6d