Fix file comments
This presumably broke in a recent refactoring: File comments do not have
the line-num attribute set, thus it returns null or empty string, which
Number converts to 0. That 0 was then used to query for the data-value
attribute on the line number elements, return null, crashing with a
"cannot get property parentElement of null" error.

The fix is to check if line-num is set, and if not query for "FILE".

Change-Id: I3335c5c8cba101ebe353954bb11ddceaad1d86ce

Former-commit-id: 6f15c45498e0e4c137f2d6f7002634cb9f5c3ce4