Move comment change handling to gr-diff-host
These three events are just used to sync the model to reflect the changes. As
nothing in gr-diff and descendents needs the comment model anymore, we can
remove that property and instead do the updating in gr-diff-host, where we
still need the comments model property to be kept in sync.

This change removes the last dependency on the comment model from gr-diff and
descendents, so they can now be reused indepedently of Gerrit's comment model.

Change-Id: I9549ea640688006c099493c2156170d706a1c65c

Former-commit-id: 5387636e8b098408ac7837a8e68a95a0dc71e9b5