Move hover handling out of gr-diff-comment
The only code that is interested in hover events from gr-comment-diff is
in gr-diff-highlight, which is used only by gr-diff. For these reasons,
it makes more sense for gr-diff (or gr-diff-highlight) to subscribe to
the native hover events of the comment threads associated with it, and
not require gr-comment to expose any custom events for that. This also
has the benefit that it reduces the requirements on thread elements:
They can just fire regular mouseenter and mouseleave events like any
DOMElement would. This makes it easier to use gr-diff with different
comment thread element implementations.

I used this change to also make the unit tests for this more meaningful
and not stub out all the interesting parts.

Change-Id: I1a969bb6a7092bc433039662cd2034dd7141e4ca

Former-commit-id: e5c377cf7a8486b755f5db37a14cf50e42283fda