Remove ReviewDb DataSource and schema setup
The goal of this change is to be able to start Gerrit without connecting
to a database. This entails removing all the DataSource bindings and
dealing with the fallout from that. This fallout includes deleting all
the schema upgrade code, which depends on the ReviewDb being a JDBC
store.

It also entails considerable changes to the bindings in low-level
injectors. The ideal future state will be to just eliminate the lowest
level dbInjector, but for now, the approach is just poke things with a
stick until tests pass.

The ReviewDb class definitely sticks around for now. However, without a
database configuration, it is never actually implemented by the gwtorm
machinery. In terms of the NoteDb wrappers, there is no longer a
concrete ReviewDb at the lowest level of wrappers. The deepest wrapped
ReviewDb is now NoChangesReviewDb, which throws
UnsupportedOperationException for every method which previously
delegated to the underlying concrete ReviewDb. (It continues to be a
no-op for actual Change-related accesses.)

Programs that depend directly on the database implementation, such as
`gsql`, now fail. Removing them is left as (near) future cleanup.

Change-Id: I6963ee23aff0f2f5c351db0377ad1256e46d9076

Former-commit-id: 3d1f85c98afe52451bac008eda5b7cda0752c090