Remove underlying ReviewDb from NoteDb wrappers
Replace the SchemaVersion access with a method that throws a runtime
exception. As a result, ReviewDb schema upgrades are no longer
supported, and we need to remove the ReviewDbSchemaVersionCheck class in
order to make the tests pass. The rest of the ReviewDb schema upgrade
machinery is dead and/or broken at this point, and will be removed in a
followup, since it doesn't actively break NoteDb tests. ReviewDb tests
are expected to fail at this point, and have been disabled in CI in
advance of this change.

After that, the only remaining Accesses in ReviewDb are change-related
tables, and all installations must be using NoteDb, so there is no
chance that they can call the real underlying ReviewDbs. We were
actually calling methods during setup like getRelationID() so we could
properly return values from the delegates, but those are not actually
called within Gerrit.

Change-Id: Ifd9bf89ea3e316aceefa83de17812b615a9b403c

Former-commit-id: b762ecc1694fe6454e63125ec21aa67589903765