Remove InitTestCase
This inherited LocalDiskRepositoryTestCase, which means it can't inherit
GerritBaseTests. (Lack of multiple inheritance is one reason why an
abstract base test class is not a great pattern to begin with, of
course.) It was doing a weird thing anyway: putting a site directory
inside the git repository created by the test case. In other words, it
was a convenient way of creating a temporary directory. In the one
actual subclass, use the more idiomatic TemporaryFolder rule instead.

Change-Id: I2b22452b53828502f82b6e522976e13e7b3227ad

Former-commit-id: eb4977f3808ab09d9a5ef760918f1a85c4adb9a6