AbstractDaemonTest: Use TemporaryFolder instead of TempFileUtil
A single instance constructed as a ClassRule is passed to
GerritServer#initAndStart. Passing the rule instance into this method is
a clear signal that the caller is responsible for cleanup, which should
just be automatic.

Compared to the previous implementation, temp site paths for failed
tests and the common server may stick around a little longer, but they
are still cleaned up when the test class finishes.

Change-Id: Iaee7961315c6efc654a881bc4b14798a668d84cb

Former-commit-id: 8a4098a30485384fcbcb675f096a252d39808307