Disable gerrit.enableReverseDnsLookup by default
Using reverse DNS lookup by default can cause unexpected performance
issues when writing the ref log, because the hostname of the user's
client is looked up on every ref log creation. For example this can
add a noticable delay (around 5 seconds) when adding a code review label,
due to the ref update of the notedb metadata refs in git.

Change the default of gerrit.enableReverseDnsLookup to false so that
DNS lookup is by default disabled.

Reported-by: Roman Karlstetter <Roman.Karlstetter@ifta.com>
Change-Id: I68af8e545fc54ea0c3819454d81bb4c598f65e93

Former-commit-id: 793525fd15812ddb7432b8e295dd073d29d998f9