Add a --no-limit option to the ssh and rest query commands
This option will return all possible results and overrides the default
limit and any limit specified using "limit" query operator. This is
useful for applications that need the entire result set with a single
query. Getting data with a single query improves the accuracy since
there is less chance of the data being modified between calls.

Change-Id: Idc4dd5799195110c580c7e6d88c0a6e8c6734d6c

Former-commit-id: b186cd759f2ab7f8050b032956a3dd01b3034a13