Add test for pushing a series of changes to non-existing master branch
Pushing changes for review requires that the target branch exists, but
there is an exception for the branch to which HEAD in the remote
repository points to. Pushing for review to this branch is allowed even
if this branch does not exist yet. The new test verifies that pushing a
series of changes to such a non-existing branch works.

Change-Id: I29d4db2691b52874bee59f0fe7bfa6649455181c
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 1ea6b5f940f910805ab77263c36775804402d666