Bazel: Update rules_closure to make it forward compatible
Gerrit cannot be built any more with upcoming Bazel 0.21 release.
That's because rules_closure depends on outdated protobuf version,
that uses some Bazel features that were discontinued in upcoming
Bazel 0.21 release.

Update rules_closure, to update protobuf to v3.6.1.3 so that we
can upgrade Bazel and build gerrit without passing incompatible
option:

  --incompatible_package_name_is_a_function=false

Another reason to upgrade urgently is because Bazel team is
maintaining Gerrit job on the CI infrastructure and this job is
broken on Bazel master.

Bug: Issue 10112
Change-Id: Icf94c3ecaeff8fd1466620cd04673a781804d21b

Former-commit-id: d003dbfbe32e75561d2ea159dfd45e843a456fd0