Fix pushing a series of changes to a non-existing branch with createNewChangeForAllNotInTarget
Pushing changes for review requires that the target branch exists, but
there is an exception for the branch to which HEAD in the remote
repository points to and for refs/meta/config. Pushing for review to
these branches is allowed even if the branch does not exist yet. This
didn't work if createNewChangeForAllNotInTarget was true because in
this case we tried to read the non-existing target branch as base and
as result the push was rejected. If the target branch doesn't exist yet
all pushed commits should always result in new changes and the setting
of createNewChangeForAllNotInTarget shouldn't matter.

Change-Id: I19a0443929d3774619a2682352fd37bca3eac321
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 2b14a2255437898dac6da4264e422c6f6ca3e244