SshCommandsIT: Add coverage of the "kill" and "ps" commands
The test was only covering commands that are under the "gerrit" root,
thus the "kill" and "ps" commands were not included. As a result of
lacking test coveraage of the "kill" command, the problem reported in
issue 10143 was overlooked.

Extend the test to also be able to cover commands that are outside of
the "gerrit" root.

ImmutableMap.of only works with up to 5 entries, so rework generation
of the list of commands to use ImmutableMap.Builder.

Coverage of the "scp" and "suexec" commands is still lacking because
the "scp" command doesn't seem to have a "--help" option and just hangs
when executed, and the "suexec" command requires additional setup.

Bug: Issue 10143
Change-Id: I0add49d2c2021d6c2961a3a212704bd5c24da17f

Former-commit-id: 92bd925d719e37019de798a4659cfeb0a7b1ff15