Bazel: Clean up package visibility settings
For packages having a default visiblity defined, it is redundant to also
set the same visibility per package, so remove those.

For packages that only have one rule, and its visibility differs from the
default visibility, remove the default visibility.

Also clean up wrapping of some of the default definitions.

Change-Id: I9e81c3f724b4ffde7a652b485d33c650866ad122

Former-commit-id: 1fa4bcfb6e1782a6ca240a404d7dcc1fa35cecea