Add converter for ChangeMessage protobuf messages
The use of ProtobufCodec requires that value classes need to be
annotated with @Column, which won't be possible as soon as we have
removed gwtorm. Hence, provide a hand-written converter for
ChangeMessage protobuf messages.

As protobuf ChangeMessage are currently used in caches and we
don't want to invalidate those, we have to ensure binary compatibility.
Prove that the new converter generates binary compatible results via
dedicated tests. Those tests will be removed after this change.

Change-Id: I7b83152597e7b3d8d4f017a07de37b1760bf3dbc

Former-commit-id: 423b9898253de779744abe792d33c6045fe72390