Remove invocation of NoteDbChecker and NoteDbMode.CHECK
NoteDbChecker was invoked after tests when GERRIT_NOTEDB=CHECK, to
ensure that data in ReviewDb and NoteDb is consistent. Since ReviewDb is
gone this is no longer needed.

NoteDbChecker cannot be removed yet since it's still used by
ChangeRebuilderIT (which will be removed once all references for
ChangeRebuilder are gone).

Change-Id: I7e413b236edb642b1aed9a8e096781098e27d286
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 30f65ceab5ce57c126215246abeab8b9fa2ae93d