Add converter for PatchSetApproval protobuf messages
The use of ProtobufCodec requires that value classes need to be
annotated with @Column, which won't be possible as soon as we have
removed gwtorm. Hence, provide a hand-written converter for
ChangeMessage protobuf messages.

As protobuf PatchSetApprovals are currently used in caches and indices
and we don't want to invalidate those, we have to ensure binary
compatibility. Prove that the new converter generates binary compatible
results via dedicated tests. Those tests will be removed after this
change.

Change-Id: I6b667caee7174c81e2892ca52ea2dc75aead9847

Former-commit-id: 1bb185ab5e505622e354da59f73aee3c40cde618