Merge NoteDbBatchUpdate into BatchUpdate
The abstract class was originally extracted to support different
implementations for ReviewDb and NoteDb. Now that there is only one
implementation, the abstract class split is just confusing.

Leaving Context and related types as interfaces is still fine, since
these types are used widely in calling code, so hiding the details of
the implementation hierarchy is warranted.

Change-Id: I394fb49f41dff5d4d3e698eb9735c6d494a35637

Former-commit-id: e51169569fd8c488762141d79a183a7f0d94d470