Remove InitNoteDb
This init step sets the config for a fully-enabled NoteDb site. In 3.0,
this is no longer necessary, as this config is not used for the serving
path. The one exception is in
NoteDbSchemaUpdater#checkNoteDbConfigFor216(), which is using the
presence of the config as a signal that the site was previously upgraded
from 2.16. From that perspective, this init step was actually
counterproductive, since it was making a 3.0 site look more like a 2.16
site than necessary.

For now, we leave the configuration in place if it was already there. A
future init step, perhaps in 3.1, can clean it up if we think it's
worthwhile.

Change-Id: I81ce2ec093e4aa1412abc3a31ffb08858be01a1a

Former-commit-id: 1724144c3164e992a02b4ba9c4c96bdd0045f8ec