Remove NotesMigration from AbstractChangeNotes.Args
Eliminate non-NoteDb conditions from subclasses. The only remaining
thing that's not quite related to the NoteDb migration is the
failOnLoadForTest mechanism, which was wedged into the NotesMigration
since it was a convenient piece of global state. Replace this with an
equally ugly piece of global state, a simple AtomicBoolean, punting for
now on a better approach.

Change-Id: I186744413ef9b263c5c6cabe19c3e5297230974f

Former-commit-id: 300af25bef281a595f3b7445afefa112d858d610