GroupApi: Add method variants with List<String>
The API includes several methods that take a vararg "String ...", which
then gets converted to a List<String> in the implementation class.

Client code that already has the input as a List<String> has to first
convert it to an array to pass to the API, which then just converts it
back to a List.

Add extra methods on the API to allow clients to pass the List<String>
directly. Replace the existing vararg methods with defaults that defer
the call to the new one. In the implementation class, implement the new
methods with List<String> and remove the conversion from array.

Change-Id: I878421e76b9bb57ee70614e6ab444336e22b8668

Former-commit-id: 0f758d82d92b1b48123a2600f32dc6f9d5b514c1