Return tags on full ref evaluation for users that have READ on refs/*
We are working on removing the shortcut that prevents Gerrit from
performing a full ref filtering if users have READ on refs/*. Gerrit's
ACLs are based on refs, not on tags. Tags are being returned if they are
reachable by any of the visible refs.

That poses the problem of what we do with orphaned tags. Due to the
shortcut, users were able to see orphaned tags in the past. In general
it seems like a sane assumption to make that when one has READ on refs/*
they can see all the tags as well. This is what this commit implements.

We considered adding a new ACL namespace (refs/tags/*) but that would
add another layer of complexity to an already hard to understand ACL
system. On top, it would make it easy for users to make code accessible
by adding a tag to a commit on a hidden ref when the administrator
granted READ on refs/tags/* to a broader audience.

Change-Id: I8d9f6a54228ed5650ded6a285c1afbf9525511e6

Former-commit-id: c7ba2ccd6f8d6a061e75c09b6eb180843aece872