NoteDbUpdateManager: Remove stageAndApplyDelta method
This method combines the stage method and applying a delta to the
NoteDbChangeState in the Change. The state in the change is only used to
store in ReviewDb, so this is no longer applicable.

The only caller is in ChangeUpdate#commit, which can just be removed, as
staging will still happen automatically during the subsequent call to
execute.

This method was the only usage of the Result class, so remove that as
well.

Change-Id: I08dbc14c801caf62c467915fef3547b744ef7bee

Former-commit-id: d9b8365dd991c012ec2326e99430f1595adefbd6