Fix NoteDb small tests to not depend on PrimaryStorage
The shouldExist argument to ChangeNotes allows loading a ChangeNotes for
a change that is yet to be created. It was incorrect for the updates
created by TestChanges#newUpdate to always set shouldExist=true; the
reality is that some tests require shouldExist=false, when creating the
first update for a new change. It was an accident that this worked at
all, since it was depending on the codepath where
PrimaryStorage=REVIEW_DB, which was for writing incomplete change
metadata to NoteDb during a migration. This migration code needs to go
away, so we can no longer rely on this accident.

Tweak the factory methods in AbstractChangeNotesTest to explicitly
specify that some updates are for new changes.

Change-Id: Ib6d30014814990ff5481a6bdf355c4ad63f66b85

Former-commit-id: 9d0ab2f4ec311c7e618bf553e108f9cd2f9cf316